Update to resolve comments on PR#2894.
Major change: Ensures that the new methods in interfaces do not break for older clients.

Minor changes: remove double cast; remove dependency on Truth assertions; remove unnecessary logic in Mutations::isNaN

+++ google-cloud-spanner/src/test/java/com/google/cloud/spanner/RandomResultSetGenerator.java
@@ -144,7 +144,7 @@ public class RandomResultSetGenerator {
           builder.setNumberValue(random.nextDouble());
           break;
         case FLOAT32:
-          builder.setNumberValue((double) random.nextFloat());
+          builder.setNumberValue(random.nextFloat());
           break;
         case INT64:
           builder.setStringValue(String.valueOf(random.nextLong()));
